SNR Enhancement for Brillouin Distributed Optical Fiber Sensors Based on Asynchronous Control
Authors: Peixuan Zhang,Biwei Wang,Yanan Yang,Abul Azad,Kuo Luo,Kuanglu YU,Changyuan Yu,Chao Lu

Abstract: We propose the asynchronous control of anisotropic diffusion (AD) algorithm, and such asynchronous anisotropic diffusion (AAD) algorithm is demonstrated experimentally to reduce noise from the sensing signals obtained from Brillouin distributed optical fiber sensors. The performance of the proposed AAD algorithm is analyzed in detail for various experimental conditions and compared with that of block-matching and 3D filtering, two-dimensional wavelet denoising, AD, and non-local means algorithms. Some key factors of the proposed algorithm, such as the effect of convolution kernel size on the performance of AD algorithms, the effect of low sampling point number (SPN) on the quality of Brillouin frequency shift and the selection of diffusion thresholds are analyzed and discussed with experimental results. The experimental results validate that the AAD algorithm can provide better root-mean-square error (RMSE) and spatial resolution (SR) than the other four algorithms, especially for higher signal-to-noise ratio (SNR) improvement and higher SPNs. For lower SPNs, the performance of AAD is also not inferior to the RMSE performance of NLM and AD. The runtime of the AAD algorithm is also quite low. Moreover, the proposed algorithm offers the best SR performance as compared to other noise reduction algorithms used in this study. Thus, the proposed AAD algorithm can be an effective candidate to improve the measurement accuracy of Brillouin distributed optical fiber sensors.